backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
django-filter==23.5
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
Pillow==10.1.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
python-dateutil==2.8.2
pytz==2023.3
